\name{Dataset 2}
\alias{met}
\alias{Gen}
\alias{Obs}
\title{
  Multi-environmental trial
}
\description{
Data collected from the SoyNAM population in Indiana 2013-2015.
}
\usage{
  data(met)
}
\details{
Data provided by Rainey Lab for Soybean Breeding and Genetics, Purdue University. Genotypic matrix (\eqn{Gen}) have 4240 SNPs. The data frame \eqn{Obs} contains the soybean id (\eqn{ID}), environment (\eqn{Year}), field location (\eqn{Block,Row,Col}) and three phenotypes: grain yield (\eqn{YLD}), days to maturity (\eqn{DTM}) and average canopy closure (\eqn{ACC}).
}
\author{
  Alencar Xavier, Ben Hall and Katy Rainey
}
\references{

Xavier, A., Hall, B., Hearst, A.A., Cherkauer, K.A. and Rainey, K.M., 2017. Genetic Architecture of Phenomic-Enabled Canopy Coverage in Glycine max. Genetics, 206(2), pp.1081-1089.

}