\name{NAM-package}
\alias{NAM-package}
\alias{NAM}
\docType{package}
\title{
Nested Association Mapping analysis
}
\description{
The NAM package was designed for association studies in nested association mapping (NAM) panels, but it also can handle biparental and random populations. It includes functions for genome-wide associations mapping of multiple populations, marker quality control, solving mixed models and finding variance components through REML.
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.0.2.1 \cr
Date: \tab 2015-23-01\cr
License: \tab GPL-3 \cr
}

}
\author{
Alencar Xavier, William Muir, Katy Rainey, Tiago Pimenta, Qishan Wang, Shizhong Xu

Maintainer: Alencar Xavier <xaviera@purdue.edu>
}

\keyword{ NAM }
\keyword{ BLUP }
\keyword{ GWAS }
\keyword{ REML }
\seealso{
Functions: blup, reml, Fst, gwas, manhattan, reference, snpQC
}