\name{Internals}
\alias{calcSize}
\alias{funI}
\alias{funX}
\alias{inputRow}
\alias{RcppExports}
\alias{timesMatrix}
\alias{timesVec}
\alias{SAMP}
\alias{.Random.seed}
\title{
  Internal functions
}
\description{
  Functions written in C++ to speed up \eqn{gwas} and \eqn{gibbs}.
}
\author{
  Alencar Xavier and Tiago Pimenta
}