\name{NAM-package}
\alias{NAM-package}
\alias{NAM}
\docType{package}
\title{
Nested Association Mapping
}
\description{
Designed for association studies in nested association mapping (NAM) panels, also handling experimental and random panels. The method is described by Xavier et al. (2015) <doi:10.1093/bioinformatics/btv448>. It includes tools for genome-wide associations of multiple populations, marker quality control, population genetics analysis, genome-wide prediction, solving mixed models and finding variance components through likelihood and Bayesian methods.
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.7.3 \cr
Date: \tab 2019-09-22\cr
License: \tab GPL-3 \cr
}

}
\author{
Alencar Xavier, William Muir, Katy Rainey, Shizhong Xu.

Maintainer: Alencar Xavier <alenxav@gmail.com>
}

\keyword{ NAM }
\keyword{ BLUP }
\keyword{ GWAS }
\keyword{ REML }
\keyword{ BGS }
\seealso{
Package include functions to perform association analysis (gwas, gwas2, gwas3), meta-gwas (gwasGE, meta3), solve mixed models (gibbs, reml), genomic prediction (wgr, gmm, press), populations genetic analysis (Fst, Gdist, LD), data quality control (snpQC, snpH2, markov), spatial analysis (covar, SPC, SPM), build kinships (GRM, GAU, PedMat) and more.
}
