\name{NAM-package}
\alias{NAM-package}
\alias{NAM}
\docType{package}
\title{
Nested Association Mapping
}
\description{
Designed for association studies in nested association mapping (NAM) panels, also handling experimental and random panels. It includes tools for genome-wide associations of multiple populations, marker quality control, population genetics analysis, genome-wide prediction, solving mixed models and finding variance components through likelihood and Bayesian methods.
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.5.4 \cr
Date: \tab 2017-10-17\cr
License: \tab GPL-3 \cr
}

}
\author{
Alencar Xavier, William Muir, Katy Rainey, Shizhong Xu.

Maintainer: Alencar Xavier <alenxav@gmail.com>
}

\keyword{ NAM }
\keyword{ BLUP }
\keyword{ GWAS }
\keyword{ REML }
\keyword{ BGS }
\seealso{
Package include functions to perform association analysis (gwas, gwas2, gwas3), meta-gwas (gwasGE, meta3), solve mixed models (gibbs, reml), genomic prediction (wgr, gmm), populations genetic analysis (Fst, Gdist, LD), data quality control (snpQC, snpH2, markov) and more.
}