\name{NAM-package}
\alias{NAM-package}
\alias{NAM}
\docType{package}
\title{
Nested Association Mapping analysis
}
\description{
NAM, biparental and random populations toolbox. The NAM package includes functions for associations mapping of multiple populations (Wang 2015), marker quality control (Rutkoski et al 2013), solving mixed models and finding variance components through REML (VanRaden 2008, Kang et al 2008, Lee et al 2006).
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.0.2 \cr
Date: \tab 2015-14-01\cr
License: \tab GPL-3 \cr
}

}
\author{
Alencar Xavier, William Muir, Katy Rainey, Tiago Pimenta, Qishan Wang, Shizhong Xu

Maintainer: Alencar Xavier <xaviera@purdue.edu>
}
\references{
Lee, S. H., & van der Werf, J. H. (2006). An efficient variance component approach implementing an average information REML suitable for combined LD and linkage mapping with a general complex pedigree. Genetics Selection Evolution, 38(1), 1-19.

Rutkoski, J. E., Poland, J., Jannink, J. L., & Sorrells, M. E. (2013). Imputation of unordered markers and the impact on genomic selection accuracy. G3: Genes| Genomes| Genetics, 3(3), 427-439.

Wang, Q. An Empirical Bayes Method for Genome-Wide Association Studies. W799/Statistical Genomics. PAG XXXII. 2015.
}

\keyword{ NAM }
\keyword{ BLUP }
\keyword{ GWAS }
\keyword{ REML }
\seealso{
Functions: blup, reml, Fst, gwas, manhattan, reference, snpQC
}