\name{NAM-package}
\alias{NAM-package}
\alias{NAM}
\docType{package}
\title{
Nested Association Mapping Analysis
}
\description{
Designed for association studies in nested association mapping (NAM) panels, also handling biparental and random panels. It includes functions for genome-wide associations mapping of multiple populations, marker quality control, solving mixed models and finding variance components through REML and Gibbs sampling.
}
\details{
\tabular{ll}{
Package: \tab NAM\cr
Type: \tab Package\cr
Version: \tab 1.2 \cr
Date: \tab 2015-04-07\cr
License: \tab GPL-3 \cr
}

}
\author{
Alencar Xavier, William Muir, Katy Rainey, Tiago Pimenta, Shizhong Xu

Maintainer: Alencar Xavier <xaviera@purdue.edu>
}

\keyword{ NAM }
\keyword{ BLUP }
\keyword{ GWAS }
\keyword{ REML }
\keyword{ BGS }
\seealso{
Functions: gibbs, reml, Fst, gwas/gwas2, reference, snpQC, snpH2
}